+++ epa/src/main/kotlin/moritz/lindner/masterarbeit/epa/api/EpaService.kt
@@ -1,10 +1,14 @@
 package moritz.lindner.masterarbeit.epa.api
 
+import io.github.oshai.kotlinlogging.KotlinLogging
 import moritz.lindner.masterarbeit.epa.ExtendedPrefixAutomaton
-import moritz.lindner.masterarbeit.epa.features.dot.DotExport
+import moritz.lindner.masterarbeit.epa.features.filter.EpaFilter
+import moritz.lindner.masterarbeit.epa.features.layout.TreeLayout
+import moritz.lindner.masterarbeit.epa.features.layout.factory.LayoutConfig
+import moritz.lindner.masterarbeit.epa.features.layout.factory.LayoutFactory
+import moritz.lindner.masterarbeit.epa.features.layout.tree.EpaToTree
 import moritz.lindner.masterarbeit.epa.features.statistics.Statistics
 import moritz.lindner.masterarbeit.epa.features.statistics.StatisticsVisitor
-import moritz.lindner.masterarbeit.epa.features.filter.EpaFilter
 
 class EpaService<T : Comparable<T>> {
 
@@ -17,24 +21,26 @@ class EpaService<T : Comparable<T>> {
     fun applyFilters(epa: ExtendedPrefixAutomaton<T>, filters: List<EpaFilter<T>>): ExtendedPrefixAutomaton<T> {
         return filters.fold(epa) { acc, filter -> filter.apply(acc) }
     }
-
-    fun getPartitionInfo(epa: ExtendedPrefixAutomaton<T>): PartitionInfo {
-        return PartitionInfo(
-            partitions = epa.getAllPartitions(),
-            totalStates = epa.states.size,
-            totalActivities = epa.activities.size
-        )
-    }
-
-    fun exportToDot(epa: ExtendedPrefixAutomaton<T>): String {
-        val visitor = DotExport<T>()
-        epa.acceptDepthFirst(visitor)
-        return visitor.dot
-    }
 }
 
-data class PartitionInfo(
-    val partitions: List<Int>,
-    val totalStates: Int,
-    val totalActivities: Int
-)
+class LayoutService<T : Comparable<T>> {
+
+    private val logger = KotlinLogging.logger {}
+
+    // TODO: maybe add caching
+    fun buildLayout(
+        epa: ExtendedPrefixAutomaton<Long>,
+        layoutConfig: LayoutConfig
+    ): TreeLayout {
+        logger.info { "building tree" }
+        val treeVisitor = EpaToTree<Long>()
+        epa.copy().acceptDepthFirst(treeVisitor)
+
+        logger.info { "building tree layout" }
+        val layout = LayoutFactory.create(layoutConfig)
+
+        layout.build(treeVisitor.root)
+
+        return layout
+    }
+}
